Simplify: remove LatticeSolution
