+++ fuzzy_lambda/src/main.rs
@@ -1,6 +1,5 @@
 use fuzzy::Output;
 use fuzzy::diff_output::{Chunk, DiffOutput};
-use fuzzy::lattice_solution::LatticeSolution;
 use fuzzy::table_solution::TableSolution;
 use fuzzy::regex_question::RegexQuestion;
 use lambda_http::{run, service_fn, Body, Error, Request, Response};
@@ -56,7 +55,7 @@ async fn function_handler(event: Request) -> Result<Response<Body>, Error> {
 
     let problem = RegexQuestion { pattern_regex: args.pattern, text: args.text }.ask()?;
     let problem_core = problem.desugar();
-    let solution: TableSolution = LatticeSolution::solve(&problem_core)?;
+    let solution = TableSolution::solve(&problem_core)?;
     let output = DiffOutput::new(&solution.score(), &solution.trace());
     let body = Out { score: *solution.score(), trace: OutChunk::from(&output.chunks) };
     let body_json = serde_json::to_string(&body)?;
